+++ src/services/mongoose/index.ts
@@ -5,3 +5,4 @@ export * from "./wallets"
 export * from "./accounts"
 export * from "./rewards"
 export * from "./ln-payments"
+export * from "./payment-flow"
